+++ src/com/cloudogu/sos/pipebuildlib/stages/DoguStageModule.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.sos.pipebuildlib
+package com.cloudogu.sos.pipebuildlib.stages
 
 interface DoguStageModule {
     void register(DoguPipe pipe, StageGroup group)
