+++ src/com/cloudogu/sos/pipebuildlib/stages/MultinodeStages.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.sos.pipebuildlib
+package com.cloudogu.sos.pipebuildlib.stages
 
 class MultinodeStages implements DoguStageModule {
